Move Compose(Network|Volume)Config to top level types
This reduces dependencies from 80 to 47

Change-type: patch
